% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mutant_apply}
\alias{mutant_apply}
\title{Apply a mutation to source text.}
\usage{
mutant_apply(source, span_start, span_end, original, replacement)
}
\arguments{
\item{source}{Original R source code}

\item{span_start}{Byte offset start of the token to replace}

\item{span_end}{Byte offset end of the token to replace}

\item{original}{The original token text}

\item{replacement}{The replacement text}
}
\value{
Mutated source text, or error string
}
\description{
Apply a mutation to source text.
}
