% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mutant_scan_file}
\alias{mutant_scan_file}
\title{Scan a single R file for mutation sites. Returns JSON string.}
\usage{
mutant_scan_file(path)
}
\arguments{
\item{path}{Path to an R file}
}
\value{
JSON string of mutation sites
}
\description{
Scan a single R file for mutation sites. Returns JSON string.
}
