% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mutant_scan_package}
\alias{mutant_scan_package}
\title{Scan all R files in a package's R/ directory. Returns JSON string.}
\usage{
mutant_scan_package(path)
}
\arguments{
\item{path}{Path to the package root directory}
}
\value{
JSON string of mutation sites
}
\description{
Scan all R files in a package's R/ directory. Returns JSON string.
}
