% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mutant_scan_source}
\alias{mutant_scan_source}
\title{Scan R source text directly. Returns JSON string of mutation sites.}
\usage{
mutant_scan_source(source, file_name)
}
\arguments{
\item{source}{R source code as a string}

\item{file_name}{Name to use for the file in results}
}
\value{
JSON string of mutation sites
}
\description{
Scan R source text directly. Returns JSON string of mutation sites.
}
